Fix compilation

+++ test/unittest/readertest.cpp
@@ -417,7 +417,7 @@ TEST(Reader, ParseNumber_NormalPrecisionError) {
         a = h.actual_;
         uint64_t bias1 = e.ToBias();
         uint64_t bias2 = a.ToBias();
-        double ulp = bias1 >= bias2 ? bias1 - bias2 : bias2 - bias1;
+        double ulp = static_cast<double>(bias1 >= bias2 ? bias1 - bias2 : bias2 - bias1);
         ulpMax = std::max(ulpMax, ulp);
         ulpSum += ulp;
     }
